Slight refactoring of CSV-upload related functions.
